feat: add opengl render command & `GetSize` for framebuffer

+++ lib_renderer/include/render_pass.h
@@ -8,7 +8,7 @@ public:
   virtual void Begin() = 0;
   virtual void End() = 0;
   virtual void SetClearColor(float r, float g, float b, float a) = 0;
-  virtual void SetViewportSize(int width, int height) = 0;
+  virtual void SetViewportSize(float width, float height) = 0;
   virtual glm::mat4 GetProjectionMatrix() const = 0;
   virtual glm::mat4 GetModelMatrix() const = 0;
   virtual ~RenderPass() = default;
